% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_px_loc.R
\name{add_px_loc}
\alias{add_px_loc}
\title{Calculating coordinates in pixels}
\usage{
add_px_loc(Spots_df, xres = 0.1625, yres = 0.1625, zres = 0.25, x_m, y_m,
  z_m)
}
\arguments{
\item{Spots_df}{a dataframe with spots information and coordinates in microns}

\item{xres}{resolution in microns for the x dimension}

\item{yres}{resolution in microns for the y dimension}

\item{zres}{resolution in microns for the z dimension}

\item{x_m}{x coordinates in pixels}

\item{y_m}{y coordinates in pixels}

\item{z_m}{z coordinates in pixels}
}
\value{
dataframe with three new columns
}
\description{
Calculating coordinates in pixels
}
