% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTrack.R
\name{checkTrack}
\alias{checkTrack}
\title{Check order of Track_df
Function checks if spots source and spots target in Track_df are inverted compared to frame and invert if necessary so that source always corresponds to the smaller frame and target to the biggest one.}
\usage{
checkTrack(Tracks_df, Spots_df)
}
\arguments{
\item{Tracks_df}{a dataframe created with Tracks.as.dataframe()}

\item{Spots_df}{a dataframe created with Spots.as.dataframe()}
}
\value{
a dataframe with reordered source and targets
}
\description{
Check order of Track_df
Function checks if spots source and spots target in Track_df are inverted compared to frame and invert if necessary so that source always corresponds to the smaller frame and target to the biggest one.
}
