% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countSpot.R
\name{countSpot}
\alias{countSpot}
\title{Counting spots per timepoints}
\usage{
countSpot(Spots_df, timepoint)
}
\arguments{
\item{Spots_df}{a dataframe created with Spots.as.Dataframe}

\item{timepoint}{column of timepoints}
}
\description{
Counting spots per timepoints
}
