% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_registration.R
\name{get_registration}
\alias{get_registration}
\title{Get information relative to registration in ViewTransform node of Big Data Viewer files
Registration based on affine transforms, one or several, written as a vector of 12 indices
This vector can be transformed as a 4x4 matrix by adding 0,0,0,1 at the end of the vector and reading it by row.}
\usage{
get_registration(Registration, name)
}
\arguments{
\item{Registration}{List of registration}

\item{name}{name of the registration}
}
\description{
Get information relative to registration in ViewTransform node of Big Data Viewer files
Registration based on affine transforms, one or several, written as a vector of 12 indices
This vector can be transformed as a 4x4 matrix by adding 0,0,0,1 at the end of the vector and reading it by row.
}
