% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap2int.R
\name{heatmap2int}
\alias{heatmap2int}
\title{Conversion of continuous variable to discrete integers for visualisation in MaMuT}
\usage{
heatmap2int(Spots_df, my_var = fluoMean)
}
\arguments{
\item{my_var}{continuous variable to convert as discrete HEX, then integer readable by MaMuT}
}
\description{
Conversion of continuous variable to discrete integers for visualisation in MaMuT
}
