% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex2int.R
\name{hex2int}
\alias{hex2int}
\title{Converting HEX colors to MaMut integer color scale}
\usage{
hex2int(x)
}
\arguments{
\item{x}{an HEX color}
}
\description{
Converting HEX colors to MaMut integer color scale
}
