% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSpotNode.R
\name{makeSpotNode}
\alias{makeSpotNode}
\title{Create a spot node
To modify the mamut.xml file.}
\usage{
makeSpotNode(...)
}
\arguments{
\item{...}{A spot dataframe}
}
\value{
a character
}
\description{
see also documentation here: https://cran.r-project.org/web/packages/xml2/vignettes/modification.html
}
