% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmFluo.R
\name{mmFluo}
\alias{mmFluo}
\title{Calculate a mean and median fluorescent value for each cell}
\usage{
mmFluo(Spots_df, all_cells)
}
\arguments{
\item{Spots_df}{a dataframe containing the spots coordinates}

\item{all_cells}{fluorescent values collected in a cube of a given size around each spot location}
}
\value{
a dataframe
}
\description{
Calculate a mean and median fluorescent value for each cell
}
