% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmFluo_all.R
\name{mmFluo_all}
\alias{mmFluo_all}
\title{Calculate mean and median fluorescent values for individual cells, per timepoin and for the whole movie}
\usage{
mmFluo_all(Spots_df, all_cells)
}
\arguments{
\item{Spots_df}{a dataframe containing the spots coordinates}

\item{all_cells}{fluorescent values collected in a cube of a given size around each spot location}
}
\description{
Calculate mean and median fluorescent values for individual cells, per timepoin and for the whole movie
}
