% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmFluo_wholeMovie.R
\name{mmFluo_wholeMovie}
\alias{mmFluo_wholeMovie}
\title{Calculate a mean and median fluorescent value for the whole movie over all spots}
\usage{
mmFluo_wholeMovie(Spots_df, all_cells)
}
\arguments{
\item{Spots_df}{a dataframe containing the spots coordinates}

\item{all_cells}{fluorescent values collected in a cube of a given size around each spot location}
}
\value{
a dataframe
}
\description{
Calculate a mean and median fluorescent value for the whole movie over all spots
}
