% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyXML.R
\name{modifyXML}
\alias{modifyXML}
\title{Modify MaMuT .xml file
Update values of existing parameters and/or add the manual color parameter}
\usage{
modifyXML(MaMuTpath, newFilename = "test.xml", Spots_df)
}
\arguments{
\item{MaMuTpath}{path to an existing MaMuT .xml file}

\item{newFilename}{new name for the modified .xml}

\item{Spots_df}{spot dataframe corresponding to the MaMuT .xml file, with or without the MANUAL_COLOR variable}
}
\value{
a new .xml file
}
\description{
Modify MaMuT .xml file
Update values of existing parameters and/or add the manual color parameter
}
