% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posLoc2posWorld.R
\name{posLoc2posWorld}
\alias{posLoc2posWorld}
\title{Calculating the coordinates
N.B.: pos_world = (3D-HASHING) x (CALIBRATION) x pos_local}
\usage{
posLoc2posWorld(pos_local, affTransform)
}
\arguments{
\item{pos_local}{Coordinates in the local coordinates system of MaMuT Viewer (POSITION_X, POSITION_Y, POSITION_Z) in arbitrary units.}

\item{affTransform}{4x4 affine transform from local to world coordinates.}
}
\value{
Coordinates corresponding to the .h5 raw data pixels
}
\description{
Calculating the coordinates
N.B.: pos_world = (3D-HASHING) x (CALIBRATION) x pos_local
}
