% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ViewRegistration.R
\name{read_ViewRegistration}
\alias{read_ViewRegistration}
\title{Get information relative to registration in .xml Big Data Viewer files}
\usage{
read_ViewRegistration(XML_BDV, setup = 0)
}
\arguments{
\item{XML_BDV}{xml file generated by the Big Data Viewer (opened to create the first MaMuT annotation file)}

\item{setup}{filter on setup value. Default to setup = 0.}
}
\description{
Get information relative to registration in .xml Big Data Viewer files
}
