% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMaMuT.R
\name{readMaMuT}
\alias{readMaMuT}
\title{Read MaMuT xml file}
\usage{
readMaMuT(MaMuTpath)
}
\arguments{
\item{MaMuTpath}{path to .xml MaMuT Fiji plugin file}
}
\value{
a list of all fields from MaMuT xml file.
readMaMuT()
}
\description{
This function is a wrapper around xmlToList function to read xml files from MaMuT plugin.
}
\keyword{3D,}
\keyword{Fiji,}
\keyword{MaMuT,}
\keyword{xml}
