% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spots.and.tracks.R
\name{spots.and.tracks}
\alias{spots.and.tracks}
\title{Combining spots and tracks dataframes}
\usage{
spots.and.tracks(Spots_df, Tracks_df)
}
\arguments{
\item{Spots_df}{dataframe created with Spots.as.dataframe()}

\item{Tracks_df}{dataframe created with Tracks.as.dataframe()}
}
\value{
a dataframe
}
\description{
Combining spots and tracks dataframes
}
