% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track2plot.R
\name{track2plot}
\alias{track2plot}
\title{Plotting tracks
Plotting MaMuT tracks with custom color code}
\usage{
track2plot(track, Tracks_df, Spots_Tracks, colNode, colYaxis, colNode_discrete)
}
\arguments{
\item{track}{track name}

\item{Tracks_df}{Tracks_df}

\item{Spots_Tracks}{Spots_df with Track name}

\item{colNode}{color of the nodes, use colNode_discrete to specify if discrete or continuous}

\item{colYaxis}{display or not y axis}

\item{colNode_discrete}{discrete or continuous, default to FALSE e.g. continuous}
}
\description{
Plotting tracks
Plotting MaMuT tracks with custom color code
}
