% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tracks.as.dataframe.R
\name{Tracks.as.dataframe}
\alias{Tracks.as.dataframe}
\title{Convert MaMuT xml Tracks tags to a dataframe}
\usage{
Tracks.as.dataframe(MaMuT_XML)
}
\arguments{
\item{MaMuT_XML}{.xml MaMuT Fiji plugin file converted to list with readMaMuT()}
}
\value{
Dataframe containing the spot ID, its name, its visibility, its radius, quality, source and position (x, y, z, t).
Tracks.as.dataframe()
}
\description{
Convert MaMuT xml Tracks tags to a dataframe
}
\keyword{3D,}
\keyword{Fiji,}
\keyword{MaMuT,}
\keyword{xml}
