% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vect2mat.R
\name{vect2mat}
\alias{vect2mat}
\title{Takes a 12 coordinates vector and returns a 4x4 matrix}
\usage{
vect2mat(affine_vect)
}
\arguments{
\item{affine_vect}{affine transform vector with 12 indices. The vector store the “top” three rows for the 4x4 affine matrix, in a "row major" fashion.}
}
\value{
a 4x4 affine transform matrix, with the top three rows and a last row containing 0,0,0,1
}
\description{
Takes a 12 coordinates vector and returns a 4x4 matrix
}
\examples{
vect2mat(1:12)
}
