% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFluo.R
\name{getFluo}
\alias{getFluo}
\title{Extracting fluorescence values
Extracting fluorescence values in an .h5 file in the BigDataViewer format around given coordinates}
\usage{
getFluo(H5_path, Spots_df, x_px = x_px, y_px = y_px, z_px = z_px,
  timepoint = FRAME, cubeSize = 5)
}
\arguments{
\item{H5_path}{path to the H5 file used to create the mamut.xml file}

\item{Spots_df}{dataframe containing on each rows spots locations at which to retrieve fluorescent values}

\item{x_px}{name of the column with x coordinates in pixels in Spots_df}

\item{y_px}{name of the column with y coordinates in pixels in Spots_df}

\item{z_px}{name of the column with z coordinates in pixels in Spots_df}

\item{timepoint}{name of the column where the timepoints or frames are stored}

\item{cubeSize}{half the length of the cube side length. By default the cube is 10x10x10 pixels.}
}
\value{
a list of length equal to the number of timepoints. For each timepoint, values of fluorescence around each spot location
}
\description{
Extracting fluorescence values
Extracting fluorescence values in an .h5 file in the BigDataViewer format around given coordinates
}
\note{
Many thanks to Dr. Mike Smith for his help on this function.
}
