% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track2plot.R
\name{track2plot}
\alias{track2plot}
\title{Plotting tracks
Plotting MaMuT tracks with custom color code}
\usage{
track2plot(track, Tracks_df, Spots_Tracks, colNode, colYaxis,
  colNode_discrete, breaks)
}
\arguments{
\item{track}{track name}

\item{Tracks_df}{Tracks_df}

\item{Spots_Tracks}{Spots_df with Track name}

\item{colNode}{color of the nodes, use colNode_discrete to specify if discrete or continuous}

\item{colYaxis}{display or not y axis}

\item{colNode_discrete}{Logical. discrete scale if TRUE, continuous otherwise. Default to FALSE e.g. continuous.}

\item{breaks}{vector of levels used to plot discrete color legend with regards to colNode unique values. Only if colNode_discrete is TRUE}
}
\value{
a plot
}
\description{
Plotting tracks
Plotting MaMuT tracks with custom color code
}
\examples{
\dontrun{
track2plot(track, Tracks_df, Spots_Tracks, colNode, colYaxis, colNode_discrete, breaks)
}
}
